Restructure components into folders
