+++ src/components/modals/Extinction.css
